feat: Add nats dependency and WebSocket publishing to Publisher class
This code change adds the "nats" dependency to the package.json file and imports the "NatsConnection" type from the "nats" package in the publisher.ts file. It also adds a new private property "wsClient" to the Publisher class, which represents the WebSocket client used for publishing to WebSocket. The "publish" method is updated to publish the event data to the NATS channel as before, but now it also checks if the "wsClient" is provided and publishes the message to the WebSocket channel if it is. This addition enhances the functionality of the code by allowing for WebSocket communication in addition to NATS publishing.