Update models/index.ts with additional fields and types
This code change adds the `requestPayload` field of type `string` and the `reason` field of type `string` to the `CTypes` namespace in the `models/index.ts` file. It also adds the `payload` field of type `any` to the `CTypes.ConnectTask` interface. These additions enhance the functionality and flexibility of the models in the application.